+++ scrapyd/__main__.py
@@ -2,9 +2,26 @@ import pkgutil
 import sys
 from os.path import dirname, join
 
-from twisted.scripts.twistd import run
+from twisted.scripts import twistd
 
 import scrapyd
+
+
+class ServerOptions(twistd.ServerOptions):
+    synopsis = 'Usage: scrapyd [options]'
+    longdesc = 'Scrapyd is an application for deploying and running Scrapy spiders.'
+
+    def __init__(self):
+        super().__init__()
+        # main() always sets -n (--nodaemon) and -y (--python=).
+        self.longOpt = [opt for opt in self.longOpt if opt not in ('nodaemon', 'python=')]
+
+    @property
+    def subCommands(self):
+        return []  # remove alternatives to running txapp.py
+
+    def getUsage(self, width=None):
+        return super().getUsage(width=width)[:-11]  # remove "\nCommands:\n"
 
 
 def main():
@@ -13,7 +30,7 @@ def main():
         print(f'Scrapyd {__version__}')
     else:
         sys.argv[1:1] = ['-n', '-y', join(dirname(scrapyd.__file__), 'txapp.py')]
-        run()
+        twistd.app.run(twistd.runApp, ServerOptions)
 
 
 if __name__ == '__main__':